feat!: implement new options for trigger creation functions
